realworld/requirements.txt
absl-py==2.1.0
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
ament-cmake-test==1.3.6
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.8.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
bleach==6.1.0
builtin-interfaces==1.2.1
cachetools==5.3.2
certifi==2025.8.3
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.90
cmake==3.28.1
CocoDataset==0.1.2
comm==0.2.2
composition-interfaces==1.2.1
contourpy==1.2.1
cv-bridge==3.2.1
cycler==0.12.1
Cython==3.0.8
debugpy==1.8.2
decorator==5.2.1
deep-sort-realtime==1.3.2
deepsort==0.0.5
demo-nodes-py==0.20.5
diagnostic-msgs==4.2.3
diagnostic-updater==3.1.2
dm-haiku==0.0.14
dm-tree==0.1.9
docstring_parser==0.17.0
domain-coordinator==0.10.0
einops==0.8.1
einshape==1.0
etils==1.13.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.59.2
fqdn==1.5.1
fsspec==2023.12.2
ftfy==6.1.3
gast==0.5.4
gdown==5.0.1
geometry-msgs==4.2.3
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.10
image-geometry==3.2.1
imageio==2.33.1
immutabledict==4.2.1
importlib-metadata==7.0.1
importlib_resources==6.5.2
interactive-markers==2.3.2
ipykernel==6.29.5
ipympl==0.9.7
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
isort==4.3.21
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
jaxline==0.0.8
jedi==0.19.1
Jinja2==3.1.3
jmp==0.0.4
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.11.3
kiwisolver==1.4.9
laser-geometry==2.4.0
launch==1.0.7
launch-ros==0.19.8
launch-testing==1.0.7
launch-testing-ros==0.19.8
launch-xml==1.0.7
launch-yaml==1.0.7
libclang==16.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
Markdown==3.5.2
markdown-it-py==4.0.0
MarkupSafe==2.1.4
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.9
mediapy==1.2.4
message-filters==4.3.4
mistune==3.0.2
ml_collections==1.1.0
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
nav-msgs==4.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.24
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
octomap-msgs==2.0.0
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
opt-einsum==3.3.0
optax==0.2.5
optree==0.17.0
osrf-pycommon==2.1.4
overrides==7.7.0
packaging==24.1
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.1.0
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==4.21.12
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-trees==2.2.3
py-trees-ros==2.2.2
py-trees-ros-interfaces==2.1.0
pyarrow==21.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
Pygments==2.19.2
pyparsing==3.2.3
pyrealsense2==2.54.2.5684
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-qt-binding==1.1.2
PyYAML==6.0.2
pyzmq==26.0.3
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
qtconsole==5.5.2
QtPy==2.4.1
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.11
rcutils==5.1.4
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
resource-retriever==3.1.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rmw-dds-common==1.6.0
ros2action==0.18.11
ros2bag==0.15.13
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.8
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.13
rosbag2-py==0.15.13
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpds-py==0.19.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9
-e git+ssh://git@github.com/facebookresearch/segment-anything-2.git@82b026cd5578af78757323ab99a0b5c8dc456cff#egg=SAM_2
scikit-learn==1.1.3
scipy==1.15.3
seaborn==0.13.1
Send2Trash==1.8.3
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
simple-parsing==0.1.7
six==1.17.0
sniffio==1.3.1
sounddevice==0.4.6
sros2==0.10.5
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.13.3
tabulate==0.9.0
tapnet @ file:///home/cindy/Documents/R2S2R/tapnet
tb-nightly==2.16.0a20240127
teleop-twist-keyboard==2.4.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.1
tensorflow-datasets==4.9.9
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorflow-metadata==1.17.2
termcolor==2.4.0
terminado==0.18.1
tf2-geometry-msgs==0.25.7
tf2-kdl==0.25.9
tf2-msgs==0.25.7
tf2-py==0.25.7
tf2-ros-py==0.25.7
tf2-tools==0.25.9
tf_keras-nightly==2.16.0.dev2024012710
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toolz==1.0.0
topic-monitor==0.20.5
torch==2.8.0+cu126
torchaudio==2.8.0+cu126
torchvision==0.23.0+cu126
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
trajectory-msgs==4.2.4
transforms3d==0.4.1
triton==3.4.0
turtlesim==1.4.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.15.0
tzdata==2023.4
ultralytics==8.3.162
ultralytics-thop==2.0.14
unique-identifier-msgs==2.2.1
uri-template==1.3.0
urllib3==2.5.0
visualization-msgs==4.2.4
wcwidth==0.2.13
webcolors==24.6.0
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.14.1
yacs==0.1.8
yapf==0.40.2
zipp==3.23.0
